Rewrite User has many gift cards to exclude carted gcs
When you add a gift card to the cart, and buy it for yourself or anyone
else, before purchasing it it is actually visible on the gift cards page
and anywhere else that grabs gift cards with user.gift_cards.

This ensures that the association will exclude all gift cards on orders
with a state of balance_due. If the gift card was created by an
administrator it won't have a line_item_id so that stipulation exists as
well